cm_conf: refactor read/write_as_str
The new code structure makes it easier to add support for new item types
that aren't integers, and thus don't share the same code paths. As part of
the conversion, switch from malloc/snprintf to asprintf to simplify the
integer code path, and avoid the need to share any code between the
u32/u16 paths.